% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{climate_balance}
\alias{climate_balance}
\title{Normative climate from manifest responses}
\usage{
climate_balance(P, eps)
}
\arguments{
\item{P}{Vector of manifest responses in \eqn{\{-1, 0, +1\}}.}

\item{eps}{Positive regularization constant \eqn{\epsilon}.}
}
\value{
A single number in \eqn{(-1, 1)}.
}
\description{
Computes \eqn{S = (n^+ - n^-) / (n^+ + n^- + \epsilon)}, where \eqn{n^+}
and \eqn{n^-} count the responses equal to \eqn{+1} and \eqn{-1}.
}
\examples{
climate_balance(c(1, 1, 1, 0, 0, 0, 0, 0, 0, 0), eps = 0.1)

# As a normative pressure in a response function
f <- response_threshold(delta = 0.5, theta = 0.4, gamma = 0.9,
                        S = function(P) climate_balance(P, eps = 0.1))

}
